pyproject.toml
[project]
name = "aurin-impact-tracking-prototype"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ddgs>=9.13.1",
    "dimcli>=0.7.0",
    "dotenv>=0.9.9",
    "google-genai>=1.73.1",
    "google-search-results>=2.4.0",
    "numpy>=1.24.0",
    "openai>=2.32.0",
    "pandas>=2.0.0",
    "pdfplumber>=0.11.9",
    "plotly>=5.15.0",
    "pycountry>=23.0.0",
    "requests>=2.31.0",
    "sqlite-utils>=3.39",
    "streamlit>=1.28.0",
    "trafilatura>=2.0.0",
    "wordcloud>=1.9.6",
    "feedparser>=6.0.0",
]

[project.scripts]
capture = "cli:main"
